chore :  파일구조 정리 및 sf3d 액세스키 깃허브 액션에 반영
