+++ .github/workflows/github-actions.yml
@@ -59,5 +59,6 @@ jobs:
                     -e S3_SECRET_KEY=${{ secrets.S3_SECRET_KEY }} \
                     -e AWS_REGION=${{ secrets.S3_REGION }} \
                     -e S3_BUCKET=${{ secrets.S3_BUCKET_NAME }} \
+                    -e STABILITY_API_KEY=${{ secrets.STABILITY_API_KEY }} \
                     -t ${{ secrets.DOCKER_USERNAME }}/${{ secrets.DOCKER_REPO }}:latest
                     sudo docker image prune -f
