+++ app/api/routes/__init__.py
@@ -1,2 +1,1 @@
 from .source import *
-from .twodto3d import *
